there's an envelope missing
manta looks like it's parsing an envelope type manually instead of
defining a proto definition for it. next step is to try to synthesize
that definition based on the manta code. it appears to be as follows:

message Envelope {
    required EDemoCommands command = 1;
    required varint tick = 2;
    required varint size = 3;
    required bytes body = 4;
}

or something to that effect. Reading protobuf primitives off the wire
manually is looney-bins.